Don't fetch sensor data if !tud_hid_ready()
Moves the tud_hid_ready() check to the beginning of the hid_task()
function, so that we don't drop sensor readings if USB is not ready.

With this change, the sensor keeps accumulating the deltas until we're
ready for them.

Also changes the default CPI to half of the previous values, which
should hopefully match previous behavior now that we're not dropping
updates.

This improves the reporting rate from 250 Hz to 500 Hz, which could
affect the running average calculations used for twist-to-scroll, but
it seems to be still working OK so I'm not modifying them for now.